psicoLE/requirements.txt
Flask
Flask-SQLAlchemy
Flask-WTF
Flask-Login
WTForms-SQLAlchemy
Flask-Babel
Flask-WeasyPrint
mercadopago

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0
pytest-mock>=3.10.0
coverage>=6.0.0
